refactor: add fromTruncted
